Fixed bug #30. Can now compile with gcc using make CC=gcc python.
This involved a fix of both the makefile- all mpiccs swapped for CC variable, and an if loop which
choose whether the DMPI_ON flag is used. Also had to take two variables which are used in non-parallel mode
and parallel mode out of the #ifdef loop in wind_updates2d.c.

This now compiles for me with gcc.