Included geo.fill in all calculations of tau_sobolev.
Also made all calculations of the sobolev optical depth explicitly use
the sobolev() function. This involved changing get_ion_density to use
a general three-vector x rather than the photon pointer, declaring
a wind pointer in p_escape, and changing the second argument of
sobolev to be a three-vector x rather than a PhotPtr. The philosophy
being that you shouldn't a PhotPtr to calculate a tau, just
a position and a line pointer - the example being that we have to calculate
p_escape in cells.