Contuiend changes to implement domains.
These change made eliminated a few parameters in geo.  The changes to
hydro_import were just to get the code to compile.  proga it turns out
is not used in anything we are currently making.  Issue #173 questions
whether we should keep proga.c (and some other routines n the
distribution.