Fixed problem with zdom[ndom].zmax not being defined for cylind_var coordinats.
While cylind_var cooridinates appeared to be working previously, in the latest domain2
branch runs, they were failing because zmax was not defined.  It's a little unclear
what caused this now, but I added code to cylind_var make_grid to assure that zmax is
defined and this appears to fix the problem.