Changes primarily to obtain the correct geo.rmax for multidomain models
The philosophy of how one sets geo.rmax is still confused, as it is set
not only in setup.c but in various other places.  The interaction
between this and zdom[ndom].rmax is complex.  This version seems to work
for an sv wind plus corona, but the solution is not pretty. Also,
clumping is a global parameter.