Partial fix of routine to make template_ionloop/shell_wind example work properly.
The probleem was that in setup.c dfudge is setup up based on
geo.wind_type, but domains allows one to have multiple wind_types in the
same simulation.  However, at present set_up_dfudge makes a special case
of shell_wind.  In reality, there is a bigger issue. As currently
written there is a single number DFUDGE, which we assume is fine for the
entire grid.  In reality DFUDGE should change depending on the size of
the cell.