Modified agn.c so that the results for template_ionloop are identical in domain2 and dev
The fix here was to add back a line that had been removed from agn.c in
domain2.  It looks like this line was removed deliberately by James, as
were a number of other lines, e.g lines specifying the datatypes of functions.
But it is clear that Nick has been modifying other parts of agn.c in dev. It
is not obvious that anything in agn.c depends on domains.  It is obvious that
these need to brought into alignment.