Modifications to add a position dependent DFUDGE to cylindrical coordinates
A new variable dfudge has been added to WindPtr.  It is populated with
the global value DFUDGE initially, but changed for cylindrical
coordinates in the subroutine that sets up the coordinate system.  It
set so that dfudge is a fraction of the size of the cell in the rho or z
direction.  The fraction is controlled from python.h.  The local dfudge
can be selectively chosen.