This commit fully removes geo.wind_type from python in favor of the wind_types in domains.
The only tricky portion of this had to do witht he fact that dfudge was
in part determined by geo.wind_type.  As of now, dfudge is totally
determined by the values of geo.rmin and geo.rmax which define the inner
and outer boundaries of the wind.

While problems are not expected, eveyone should be on the lookout for
problems with stuck photons.  This would indicate the there are
situations where DFUDGE is not defined appropriately.

The problem with geo.wind_type is discussed in #212