Clean up after checking that various restarts of an old run or when reading in an old wind file works with domains.
As far as I can tell restarts work properly with single or multiple
domains.  The only functional change I made was to make sure that
when an old windsave file was read in, that the completed ionization and
spectral cycles were set to 0.

The appropriate way to restart and old run is to use the -r option, and
in this case the program only computes the cycles that it has not done
previously.