Documentation updates after checking that the yso model works as it did originally without domains.
Note that the yso model is still implemented as a single domain, even
though the wind consists of a spherical stellar wind and a KWD wind
outside of it.  It would more naturally be reimplemented as a two
component wind, but this should probably be done as part of a more
general project to create a multi-component wind.  In that case,
 the yso model could be used as a check.