Updates to spherical.c to properly accommodate domains
There were errors that would have caused the spherical coordintes to
fail if there were multiple domains.  They are now fixed, and I have
checked that the results for a single domain are the same as previously.
This points to the need, if we are going to use domains to develop a
robust way to actually test multiple domains systematically.

The only acutal changes were made to spherical.c. All else was
documatation improvment.