Continuation of attempt to use number defines or enums for choice statemetns in python and py_wind
At this point, my review and fixes for this are essentially complete.  I
have made changes only for those varaibles that were defined as enums or
with #define statements.  There are other variables that should be
similar defined, e. g. geo.scatter_mode but I have not added this.

I found that there are vestigial geo.ion_modes that probably should be
romoved.  This is described in Bug #215