Small changes to get rid of warnings when gcc-4 was used to compile python.
Most of the cases fixed were places where a variable was defined and
intialized but not actually used in python.  A few were logical
inconsistences where a bunch of if statments did not have a default.
There are still some errors.  Almost all of these are in routines that
James or Nick have been working on and I would like to give them an
option to fix these, as they know these routines better than I.