First version of routine which incorporates a capability to define wind components generically.
A new variable is requested asking for the number of wind components.
Then one is queried as to the type of wind component and the size
of the grid.  A disk atmosphere is still treated separately, though
it should be easy to eliminate this.

Although the place where one is asked to define the disk type has
been moved, these changes are not really about rearanging the inputs.