This update significantly changes its functionality of rdpar.
The main goal of these changes was to accommodate a reordering of
the parameter files, to accomodate the possibility of having
multiple .pf files, and to allow for writing comments into the output .pf files.

With the new version of rdpar, the input .pf file (or files) are read
into a structure at the beginning of the program.  The lines in this
structure are then parsed as data are read in.  When a line matches
the input keyword it is marked as USED, so a line can only be matched once.
However, each time one searches for an input keyword, the routine searches
all of the UNUSED lines from the beginning of the file.  This means that order
of the keywords in the input file does not matter, unless a keyword is
requested multiple times.

Additionally, if a keyword is missing, the program asks for that keyword
in interactive mode, but returns to non-interactive mode for the next keyword.
This should simplify the situation where new keywords are added.

A new subroutine add_par has been added. This allows one to concatenate
multiple .pf files on input. opar should be called first.  This could
allow one to add a debug.pf file, for example.

The behavior of the files produced by cpar is changed.  The output
file will be in the order that the variables are actually needed by rdint, etc.
This means that extra variables and comments in the input .pf file or
files will be cleaned out.  Even if there are multiple .pf files on input,
there is only one output file on exit.

A new subroutine rdpar_comment has been added. This puts a comment in files produces
by cpar, and is intended so that one can separate portions of the inputs in a logical
manner. For example, one might put a comment at the beginning of the description of each domain.