Modified rdpar so that a warning was not printed out unless the rdpar verbosity parameter was large
As now written, rdpar can read the .pf file multiple times and this was
generating lots of warnings, when for example two lines in a rdpar file
were interchanged.

Note: It is not completely clear that the "verbose" parameter is being
used the way we want it to be.  In particular there are statements in
rdpar that require a particular value of verbose to induce a print out.
We proably want the verbosity to increase with the value of verbose.