Change the way  "root" is constructed from the parameter file name.
Previously, the root was taken to be the input string up to the first
".".  Now it is the input string up the the first (!) instance of ".pf".

There remains a possibility of confusion if the input string is
something liek "hello.pfoo.pf.  In this case the root would end up
being "hello".

The change was made to addres issue #228

Please enter the commit message for your changes. Lines starting