Small change to inization.c to restore former functionality
In trying to conistently use -f as the only way to fix the electron
temperatures, I somehow introduced a problem which caused problems with
calculating dvds.  While this change prevents that problem, it may now
be that -f is not the only way to fix the temperature, and therefore
I may have undone the earlier fix.  So this is a quickfix to restore
functionality, and the consistency quesion needs testing