Modifications to allow recomb.c to return the full recombination rates for normal fb
At present, the outputs of a 1d thick shell model are unchanged.  I
believe this due the the fact that recomb routines are still returning
the reduced rate even though the full rate is being requested.  This may
be due to way in which we store fb values and get them back from
routines like get_fb