Corrected problem in in recomb.c associated with the number of jumps
With mulitple jumps an error in pdf.c was causing the program to exit
due to the fact that the jumps were not in ascending order.  The
actual problme was due to having jumps with the same value.  This
occurs becaue in for expample H there are mulitple levels with the
same potential.  The fix carried out in recomb.c is to sort the jumps
and then  eliminate any duplicates.