Modifications to allow cdfs to be created which are smaller than the maximum size.
If the input pdf array was not well sampled in the sense that a
significant fraction of it had near zero values than our attempt to
sample onto a uniform grid was failing because the cdf jumped by more
than the default value of 0.05 between points in the input cdf. The
solution in generating the cdf is to allow for the cdf array to have
fewer points.  In practice one needs to assure that the input array is
large enough that it properly samples the pdf.