Modified cdf_gen_from_array so that xmin and xmax did not have to be the mininum and maximum value of the input array.
As I understand Nick's changes, the main ones were to remove jumps from
pdf_gen_from array, and to replace the old scheme of subsampling so that
the one had the full resolution of the array.

He also made other changes.  These include requiring that the xmin and
xmax correspond to the min and maximum of the array instead of the
(frequency) region the calling program desired.  Being able to specify
the range independent of the array was an important capability that was
lost.  It is now restored.

Nick also tried to eliminate zeroes at the beginning and the end, but
this makes the programming more complex so I have eliminated this part
of the code.  If there is a reason it is required, it can be restored.
While having zeroes at the end or beginning of the array is probably
most common, there is no reason that there could not be a series of
zeroes in the middle of a pdf and we have to handle that situation
anyway.

This is all relevent to isssues #297 #294 and #270