First part of effort to handle in a better way photons which hit the disk or star (See #260)
At this point instead of geo.disk_illum there is a new variable
geo.absorb_reflect, which is intended to govern all processing of
photons which hit either the disk or star.  The rdpar variable has been
renamed to Surface.reflection.or.absorption.  Updates have been made to synonyms
so the old variable can still be understood.

The option of having an analytic approximation for the stellar temperature
has been moved to the disk.temperature.profile rdpar line, and to the
geo.tprofile variable.

Scattering off the star has been implemented, and the way photons
scatter off the disk has been changed so it is not specular reflection.

Star_init has been modified to calculate a revised stellar temperature,
including the back scattered radiation in the case where we want back
scattered photons to heat the star (and disk).

Lots of this needs detailed testing.