Steps toward being able to import models more generically. #309
In addition to creating a new routine import.c to read in models
I have moved some of the inputs having to do with winds from python.c
into a reanmed routine in setup, to collect the infomrmation about
domains.  Note that to avoid having to specify the dimensions in
advance one needs to actually read in the models, before one allocates
wmain.