Updated where disk_params are read in to address issue #214
The order in which the parameters are read in has been revised so that
one is not asking for unnecessary paramters.  All of the rdpar inputs
for disks have been moved in to get_disk_params.  The get_disk_param
routine has been moved into a separate file setup_disk.c