Updates to the way line photons are generated in the wind as part of effort to address issue #311
As part of the effort, elements npdf, pdf_x and pdf_y were removed from
the Plasma structure, and more generally emission.c was cleaned up.
These changes have not solved #311 yet, which is arising because for a
particular we are trying to create line photons for a band that has no
lines.