Updates to add shock_heating as a term in zero_emit. Changes to windsavewtable to record extra heating source
This version of the code is the first that actually attempts to solve
for the temperature in a model with shock heating.  It does produce
models with t_e substantially higher than t_r in the portions of the
wind where shock_heating dominates.  No macro_specific modifications
have been made.