Updated rdchoice so that minimum match for variables is restored.
As discussed in #444, Nick had made changes which changes which replaced
a minimum match criterion with an exact match criterion.  With this
update, we go back to the mimumum match criterion, but also check whethe
the minimum match is ambigouos.  If it is, then the query is reissued.
The specific issue associated with coordinate system naminc convetions
ahs also been eliminated.  Specically, the coordinate system name
spherical_polar has been replaced by polar.