Update to enforce up-to-date input files.   See #461
Python has been updated so that the old formats cause the program to generate
an output .pf file with the keywords updated, and containing warnings about
using integers instead of words for rdchoice variables.  But then the
program halts.  The user must correct the input file for the program to
run to completion.

.pf files for the standard regreession checks have been updated.