Fix issues associated with an inconsitent defination of the geometry of a vertically extended disk.
A vertically extended disk is now defined by a power law increase in the
radius out to geo.radmax, and a cylinder on the outside edge.
(Previously there was confusing in various places in the program about
whether the outside edge was defined by a cylinder or a sphere).

See #159