Bug503 restart (#510)
Fix for renormalization issues on restarts.  See #503

At this point, the code should handle both of the use cases for restarts,
the first when the program is stopped and then restarted without changing
the .pf file, and the second when the code is run to completion, the number
of spectral cycles increased, and the program rerun with the restart option.

The problem with the normalization of the created spectra has also been resolved.