Converge (#578)
* More cleanup of run_check.py and related routines

*  Fixes #577 where the old radiation temperature was not being recorded properly.
 Modifications to windsave2table to print out information about convergence

* Updates to run_check and py_error as part of the effort to provide
better diagnostics for users.

Note that because some duplicated outputs were removed from the diag
files, the lines associated with reading the convergence summary have
changed.  So other programs that parse the log files to search for this may have to be modified.  One can see how to do this in run_check.py