Updates to allow continued checking of the hydro commmands  (#589)
The old way of checking the hydro commands involved a comparison of heating and cooling rates to a stored version of the cooling rates. These change in detail whenever the random number generator changes, so no one was paying attention. The new version compares to the previous run of the regression tests, so that changes should be easier to spot.

See #586