More fixes to move the calculation of doppler shifts to the routines in frames.c
Results no longer agree precisely with dev, but there is one place in dev where we did not
follow the correct way to tranform back from the local to observer frame. Additinal
changes in resonate.c are still needed to complete pulling all of the doppler shifts
out of the individual routines.