Bug700 vert (#705) - Vertically extended disks
These modifications represent a significant update to the way vertically extended disks
are handled.    They do not affect, or at least should not affect anything having to do with flat disks.

To do this, disk_init and walls have been refactored (moved) into separate routines, and ds_to_disk for vertically extended disks has been completely rewritten.

These changes were  intended to make it easier to diagnose issues with vertically
extended disks, and ultimately to make it easier to add disks with
different temperature distributions, see e.g. #660

qdisk_init has been re-writtend so that the annulae are logarithmically spaced, and not dependent on the exact temperature distribution of the disk

ds_to_disk has been rewritten completely for vertically extended disks; it should not 
be changed for flat disks

Fixed stuff_phot so it copies all of the variables, ds had been omitted

Fixed problems in sv.c and in knigge.c so the wind models are defined sensibly for 
vertically extended disks

There are a few residual concerns that should be addressed.  Scattering off of the outside edge of a vertically extended disk is not properly handled.  In walls.c (for flat and vertically extended disks) we check whether a photon hits the central object early in the program and if it does we assume it has been reflected there.  It is not entirely obvious that the photon may not hit the disk first; more investigation of this is needed.