Bug645 repo (#708)
* Fixes to trans_phot to set ds to 0, when a photons scatters
* Fixes to walls to properly deal with hitting the star.    Removal of a supplementary
reposition routine.  

At this point, the determination within walls of whether a photon has hit the disk or star is based entirely on the position and direction of the old and proposed photon.  ds which is calculated externally to walls is not longer used.

Note: At some point in the near future, once we have used this version of walls for several weeks or months, there are "belt and suspenders" checks in Python that should be considered for removal as they simply take extra compute cycles.  The most obvious one to delete is the extra call to walls after a scatter which should never be needed if reposition has worked correctly