Merge latest dev changes into frame branch (#709)
* spherical flux

* para helper elbarged

* para helper elbarged

* Fixed rad_hydro_files problems caused by change of flux definition in spherical coords

* Fixed errors caused in rad_hydro_files by negative fluxes

* Add void * to brem_d and update CDF functions

* finished changes to address bug #665

* increased version number due to changes to windsave

* Documentation updates and rename variable to be more consistent

* Fix additional compiler warnings about function pointers

* Set a line back to an Error (was log_silent) in trans_phot

* Better labelling for regression_plots. Fixes #707

* Bug645 repo (#708)


* Fixes to trans_phot to set ds to 0, when a photons scatters
* Fixes to walls to properly deal with hitting the star.    Removal of a supplementary
reposition routine.  

At this point, the determination within walls of whether a photon has hit the disk or star is based entirely on the position and direction of the old and proposed photon.  ds which is calculated externally to walls is not longer used.

Note: At some point in the near future, once we have used this version of walls for several weeks or months, there are "belt and suspenders" checks in Python that should be considered for removal as they simply take extra compute cycles.  The most obvious one to delete is the extra call to walls after a scatter which should never be needed if reposition has worked correctly

Co-authored-by: higginbottom <nick_higginbottom@fastmail.fm>
Co-authored-by: Edward Parkinson <e.j.parkinson@soton.ac.uk>