Bug700 vert (#729)
This change  represents "final" changes (and some fixes) to address bug #700.   Specifically changes here have been made to allow for slight imprecision in calculating the distance to a vertically extended disk. It also allows for the possibility that a wind photon hits the outer edge of an extended disk.

None of the changes here affect flat disks.