Updated regression.py so that a file commands.txt can optionally be used to add command line switches to individual runs of pf files ik2 a regression directory.
Any switches can be included, but the proximate reason for this was
to allow one to run some models with full special relativity.

regrresion.py will not stop altogether if a particular modle fails,
becasue a command line switch (e.g. -gamma) is not available in an
earlier version of python. That model will fail but regression
checks will contine on other models in the regression directory

Updated regression tests; old regression tests are in regress.old