Bug739 spectra (#740)
* Remove freq_orig_loc as not needed. 

* Machinery to record spectra on the fly for disk photons and for (in the macro atom case) wind photons

* Added a new column containing the spectra of the photons
that are created in the wind.  Various plotting routines have also
been updated to reflect the changed spectrum file formats.  The column that is labeled created in the spectra
now refers only to photons generated from external sources (the central object  or the disk).

*  Improvements have been made to the detailed spectral plots for the regression_check
routine.   

These changes were made in association with issue #739