Update to extract to properly calculate aberrations (#768)
Rearranges some of the code in extract, and adds code necessary to assure that when we extract photons in a particular direction, aberration effects are properly accounted for.  Basically all that is done is a observer to local frame conversion for a dummy photon packet, so that in the cmf frame we create a photon that when transferred back to the observer frame will be going in the correct observer frame direction.