Bug819 bb (#820)
Corrects problem with reading in disk models, when read in model does not extend over the full disk, see #819.  

As revised, when one reads the temperature structure of a non-standard disk into python, it sets the outer radius of the disk to the maximum radius in the read in model.