Bug878 gaunt (#882)
Changes to prevent repeat calculations of portions of ff
emission on calls from matom.  See #878

For optically thick media calculated in macro atom mode, this change reduces the percentage of time spent on ff to well below what it was previously.  With these changes, ff is no longer a major time sink.