Eliminate gcc warnings associated with sprintf
These warnings were all asscoiatec with the possibility that sprintf
would overload the size of the targed string.  The solution was to
specify the "precision" of each of the inputs strings using %.10s,
for example to limit the string lenght to in this case 10.  The
choices of string lengths hould be checked.