Gcc11 (#913)
* Make changes to eliminate warnings an error associated with compiling under gcc11 (#873)

There are two major categories

* Global variables all need to be labelled as extern in the .h file  and then initialized i(in a c file).

In future, one should be sure to global variables declare as extern  in the relevant .h  and intialize it in the corresponding .c file, named in the case of python.h,  python_extern_init.c

* A number of sprintf statements generated warning due to the possibility that they would overrun the buffer

These warnings were all associated with the possibility that sprintf
would overload the size of the targed string.  The solution was to
specify the "precision" of each of the inputs strings using %.10s,
for example to limit the string lenght to in this case 10.  The
choices of string lengths should be checked.

Additionally there were a number of variables that were set but not used that generate warnings.

Additionally removed several .h files, which were not necessary or not being used, or in a few cases so very small that they could be moved into one of the larger files

* Set version to 85j

At the end of this process, none of the routines shows any warnings under gcc11. The results of the regression tests are unchanged from dev when compiled normally, but do differ slightly in gcc11.