Change from cartesian to spherical fluxes
I changed the fluxes that are computed in estimators_simple.c so that those are stored in spherical coordinates instead of cartesian coordinates. This avoids changing frame again in PLUTO and minimize errors due to a coarse grid. I changed the names of the structures like F_UV_x/y/z to F_UV_theta/phi/r.