Clean up issue #963
The issue was that in live or die mode, one cannot with
advanced diagnositcs, select photon by position. As it
is not clear, why one would want to do this, and since
it would require effort to fix.  The program exits and
advises the user to use extract mode instead.