Polynesia - Updates to address various issues and add capability to generate simple data files directly from chianti (#1052)
* Retrun to more reliable gsl integration

* Fixes for photon in cell w/out wind

* Update of inspect_wind

* Remove diagnostic line

* Update to deal with bug_tardis

* Add Ca to windsave2table for 1d sn comparison

* Doxygen doc for collisional rates

* Update to RedoPhot so can handle non-macro files as well

* Sphinx updates explaining  #834

Live or die and extract can differ in
models where one is extracting along
a wind cone.

* Clean up issue #963

The issue was that in live or die mode, one cannot with
advanced diagnostics, select photon by position. As it
is not clear, why one would want to do this, and since
it would require effort to fix.  The program exits and
advises the user to use extract mode instead.

* A routine to convert macro to simple atomic models

The macro models must have been made with MakeMacro

* Fix import.py routines to obtain correct rho

Windsave2table reports CMF rho's. Python
expects observer frame rho's.  This is
a simple fix in the import reoutiens to
reflect this.  See Issue #787


* Change version to 87i

* Change of ion mode for balmer test